collect python scripts into 2 folders: PythonTools and JobModule
