+++ PythonTools/DebugManager.py
