--- the_conf.yaml
@@ -1,53 +1,0 @@
-
-basic_parameters:
-  timeout: 0.4
-  host: 'ntugrid8.phys.ntu.edu.tw'
-  user: 'ltsai'
-  pkey: '/Users/noises/.ssh/toNTU8'
-cmd_templates:
-  'init': 'echo "connection established"'
-  'run': 'python3 -u main_job.py; echo "config: {prefix}", echo "additional config: {ip2}"'
-  'stop': 'exit'
-  'del': ''
-cmd_arguments:
-  prefix: default
-cmd_const_arguments:
-  ip2: ntugrid5.phys.ntu.edu.tw
-logging:
-  stdout:
-    name: out
-    file: log_stdout.txt
-    filters:
-      - indicator: running
-        threshold: 0
-        pattern: 'RUNNING'
-        filter_method: exact
-      - indicator: Type0ERROR
-        threshold: 0
-        pattern: '[running] 0'
-        filter_method: exact
-      - indicator: Type3ERROR
-        threshold: 0
-        pattern: '[running] 3'
-        filter_method: contain
-      - indicator: idle
-        threshold: 0
-        pattern: 'FINISHED'
-        filter_method: exact
-  stderr:
-    name: err
-    file: log_stderr.txt
-    filters:
-      - indicator: running
-        threshold: 0
-        pattern: 'RUNNING'
-        filter_method: exact
-      - indicator: Type0ERROR
-        threshold: 0
-        pattern: '[running] 0'
-        filter_method: exact
-      - indicator: RaiseError
-        threshold: 0
-        pattern: 'Error'
-        filter_method: contain
-
